rename utilities to load_test_robot
